+++ packages/flutter_rocket/example/lib/requests/todos.dart
@@ -5,5 +5,5 @@ const String todosEndpoint = "todos";
 
 class GetTodos {
   static Future gettodos(Todos todosModel) =>
-      Rocket.get<RocketClient>().request(todosEndpoint, model: todosModel);
+      Rocket.get<RocketClient>()!.request(todosEndpoint, model: todosModel);
 }